Refactor config system
- Archive old config system in config_2 (will be removed when new one is finished)
- Add new independent config-loader
