+++ src/main/kotlin/wtf/votebot/bot/config/ConfigBackend.kt
@@ -19,5 +19,5 @@
 
 package wtf.votebot.bot.config
 
-interface Config {
+interface ConfigBackend {
 }
