+++ src/main/kotlin/wtf/votebot/bot/config/_ConfigAnnotations.kt
@@ -1,0 +1,2 @@
+package wtf.votebot.bot.config
+
