+++ src/main/kotlin/wtf/votebot/bot/config_2/ConfigLoader.kt
@@ -19,5 +19,22 @@
 
 package wtf.votebot.bot.config
 
-interface Config {
+import wtf.votebot.bot.exceptions.StartupError
+
+class ConfigLoader(
+    envConfig: EnvConfig,
+    vaultConfig: VaultConfig?
+) : Config {
+
+    override val environment = envConfig.environment ?: ConfigDefaults.ENVIRONMENT
+    override val httpPort = envConfig.httpPort ?: ConfigDefaults.HTTP_PORT
+    override val sentryDSN = vaultConfig?.sentryDSN ?: envConfig.sentryDSN
+    override val discordToken = vaultConfig?.discordToken ?: envConfig.discordToken ?: requiredNotFound("DiscordToken")
+
+    companion object {
+        fun requiredNotFound(option: String): Nothing = throw StartupError(
+            "Could not find $option in any of the config backends. " +
+                    "Please make sure to include all required options."
+        )
+    }
 }
