+++ src/main/kotlin/wtf/votebot/bot/config_2/VaultConfig.kt
